requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
assemblyai==0.34.0
attrs==24.2.0
boto3==1.35.42
botocore==1.35.42
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
distro==1.9.0
filelock==3.16.1
frozenlist==1.4.1
grpcio==1.67.0
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jiter==0.6.1
jmespath==1.0.1
moviepy==1.0.3
multidict==6.1.0
numpy==2.1.2
openai==1.51.2
pillow==11.0.0
proglog==0.1.10
propcache==0.2.0
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyht==0.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.15.3
